Version 6.0.1.5 Done (#302)
* chore: Update navigation-compose to 2.8.0-beta04

Updated the version of navigation-compose dependency to 2.8
.0-beta04 in the version catalog.

* Pre-release 6.0.1.4.5 done

Increased versionCode to 81 and versionName to 6.0.1
.4.5.

* - Dependencies Upgrade
- Migrating Kotlin to V2
- Compose Compiler migrated to V2

* Feat: Implement JetLime for attendance calendar

This commit introduces JetLime library to enhance the attendance calendar view. It replaces the previous custom implementation with JetLime'
s `JetLimeColumn` and `JetLimeExtendedEvent` components, providing a more polished and interactive user experience.

The commit also includes minor refactoring and adjustments to the layout and styling of the attendance calendar.

* Pre-release 6.0.1.4.6 done

Increased versionCode to 82 and versionName to 6.0.1
.4.6.

* Pre-release 6.0.1.4.6 done

chore: Suppress warnings from Android Gradle plugin

This commit adds rules to the Proguard configuration files to suppress warnings generated by the Android Gradle plugin
. These warnings are related to various libraries and components used in the project.

* Version 6.0.1.5 done
chore: Upgrade dependencies and migrate to Kotlin 2.0

This commit upgrades several dependencies, including:

- AGP to 8.6.
0-rc01
- Coil to 2.7.0
- Coil Compose to 2.7.0
- Hilt to 2.52
- JetLime to 2.2.1
- Lottie Compose to 6.5.0
- Material 3 to
 1.3.0-beta05
- Navigation Compose to 2.8.0-beta07
- Paging Compose to 3.3.2
- Paging Runtime Ktx to 3.3.2

It also migrates the project to Kotlin 2.0 and updates the compile and target SDK versions to 35.

Additionally, the commit disables minification in release builds for chat, core, and syllabus modules to address potential issues.